ADM-652 [frontend] feat: split detail component
